+++ src/app/(post-login)/submission/program/[shortName]/clinical-submission/components/Header.tsx
@@ -21,12 +21,12 @@ import APPROVE_SUBMISSION_MUTATION from '@/app/gql/APPROVE_SUBMISSION_MUTATION';
 import CLEAR_CLINICAL_SUBMISSION from '@/app/gql/CLEAR_CLINICAL_SUBMISSION';
 import REOPEN_SUBMISSION_MUTATION from '@/app/gql/REOPEN_SUBMISSION_MUTATION';
 import { useAppConfigContext } from '@/app/hooks/AppProvider';
+import { useAuthContext } from '@/app/hooks/AuthProvider';
 import { useGlobalLoader } from '@/app/hooks/GlobalLoaderProvider';
 import { useToaster } from '@/app/hooks/ToastProvider';
 import useCommonToasters from '@/app/hooks/useCommonToasters';
 import { useSubmissionSystemStatus } from '@/app/hooks/useSubmissionSystemStatus';
 import { sleep } from '@/global/utils';
-import { useAuthContext } from '@/global/utils/auth';
 import { css, useTheme } from '@/lib/emotion';
 import { useMutation } from '@apollo/client';
 import { Button, TitleBar, Link as UIKitLink } from '@icgc-argo/uikit';
